requirements.txt
pdfplumber>=0.7.6
Pillow>=9.0.0
scikit-learn>=1.0.0
numpy>=1.21.0
